Move to NotificationController
