+++ app.js
@@ -2,7 +2,7 @@ import express from 'express';
 import mongoose from 'mongoose';
 import { configDotenv } from 'dotenv';
 import run from './config/testConnection.js';
-import { sendDueNotifications } from './services/notifications.js';
+import { sendDueNotifications } from './controllers/notificationController.js';
 configDotenv();
 
 const port = process.env.PORT || 3300;
